change Rect and Size to use i32 instead of u32

+++ crates/ayed-core/src/panels/line_numbers.rs
@@ -14,7 +14,7 @@ pub struct LineNumbers {
 }
 
 impl LineNumbers {
-    const RIGHT_PAD_LEN: u32 = 2;
+    const RIGHT_PAD_LEN: i32 = 2;
 
     pub fn rect(&self) -> Rect {
         self.rect
@@ -24,13 +24,13 @@ impl LineNumbers {
         self.rect = rect;
     }
 
-    pub fn required_width(&self, ctx: &RenderPanelContext) -> u32 {
+    pub fn required_width(&self, ctx: &RenderPanelContext) -> i32 {
         let Some(buffer_handle) = ctx.state.active_editor_buffer(&ctx.resources) else {
             return 2;
         };
         let max_line = ctx.resources.buffers.get(buffer_handle).line_count();
-        const LEFT_PAD_LEN: u32 = 1;
-        let width = (max_line.ilog10() + 1) + LEFT_PAD_LEN + Self::RIGHT_PAD_LEN;
+        const LEFT_PAD_LEN: i32 = 1;
+        let width = ((max_line.ilog10() as i32) + 1) + LEFT_PAD_LEN + Self::RIGHT_PAD_LEN;
         width
     }
